+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/structure/ProtobufScope.kt
@@ -2,6 +2,7 @@ package io.kanro.idea.plugin.protobuf.lang.psi.primitive.structure
 
 import com.intellij.psi.util.QualifiedName
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufReservedName
+import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufReservedStatement
 import io.kanro.idea.plugin.protobuf.lang.psi.findChildren
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.stratify.ProtobufBodyOwner
 
@@ -10,9 +11,11 @@ interface ProtobufScope : ProtobufScopeItemContainer, ProtobufScopeItem {
 
     fun reservedNames(): Array<ProtobufReservedName> {
         return if (this is ProtobufBodyOwner) {
-            this.body()?.findChildren() ?: arrayOf()
+            this.body()?.findChildren<ProtobufReservedStatement>() ?: arrayOf()
         } else {
             findChildren()
-        }
+        }.flatMap {
+            it.reservedNameList
+        }.toTypedArray()
     }
 }